Moved more into parent class, added javadocs for transactionId
